line-item move to type guard
